Add code coverage combination and reporting (#1674)
SUMMARY:
Update code coverage support to include results in the job summaries, as
well as combine the results from the multiple jobs in the
"test-checks.yaml" workflow.

Some open questions:
* Should the reporting in the summary skip empty files (files with no
Python code/statements)?
  * Current: yes
* Should the reporting use different types of sorting (e.g., by
coverage)?
  * Current: file path/name (tool default)
* Should the GitHub jobs have coverage summaries for intermediate jobs
or just the combined results?
  * Current: all (though it does seem more verbose than is useful)
* Note: The coverage result artifacts will still be attached to the jobs
even without summaries, they are currently required for the job that
combines the results.
* Should coverage be enabled by default for eithe of the workflows (here
or another PR)?
* Current: no (adds ~1 minute to base test run wall time and ~6-7
minutes to transformers run time)

TEST PLAN:
For examples, check these runs and the summary for each job in the run:

* base:
https://github.com/vllm-project/llm-compressor/actions/runs/16471546227
(49% combined coverage)
* transformers:
https://github.com/vllm-project/llm-compressor/actions/runs/16472609090
(65% coverage)

Signed-off-by: Domenic Barbuzzi <domenic@neuralmagic.com>